+++ decidim-sortitions/spec/queries/decidim/sortitions/admin/participatory_space_proposals_spec.rb
@@ -85,6 +85,16 @@ module Decidim
             expect(described_class.for(sortition)).not_to include(*proposals)
           end
         end
+
+        context "when filtering drafts proposals" do
+          let(:proposals) do
+            create_list(:proposal, 10, :draft, component: proposal_component, created_at: request_timestamp - 10.days)
+          end
+
+          it "do not return draft proposals" do
+            expect(described_class.for(sortition)).not_to include(*proposals)
+          end
+        end
       end
     end
   end